feat: Makefile

+++ src/analytics/exec_query.py
@@ -28,7 +28,7 @@ def date_range(start, stop, monthly=False):
     return dates
 
 
-def exec_query(table, db_origin, db_target,dt_start, dt_stop, monthly):
+def exec_query(table, db_origin, db_target,dt_start, dt_stop, monthly, mode='append'):
 
     engine_app = sqlalchemy.create_engine(f"sqlite:///../../data/{db_origin}/database.db")
     engine_analytical = sqlalchemy.create_engine(f"sqlite:///../../data/{db_target}/database.db")
@@ -37,19 +37,18 @@ def exec_query(table, db_origin, db_target,dt_start, dt_stop, monthly):
     dates = date_range(dt_start, dt_stop, monthly)
 
     for i in tqdm(dates):
+        if mode == 'append':
+            with engine_analytical.connect() as con:
+                try:
+                    query_delete = f"DELETE FROM {table} WHERE dtRef = date('{i}', '-1 day')"
+                    con.execute(sqlalchemy.text(query_delete))
+                    con.commit()
+                except Exception as err:
+                    print(err)
 
-        with engine_analytical.connect() as con:
-            try:
-                query_delete = f"DELETE FROM {table} WHERE dtRef = date('{i}', '-1 day')"
-                con.execute(sqlalchemy.text(query_delete))
-                con.commit()
-            except Exception as err:
-                print(err)
-
-        # print(i)
         query_format = query.format(date=i)
         df = pd.read_sql(query_format, engine_app)
-        df.to_sql(table, engine_analytical, index=False, if_exists="append")
+        df.to_sql(table, engine_analytical, index=False, if_exists=mode)
 
 def main():
     parser = argparse.ArgumentParser()
@@ -63,6 +62,7 @@ def main():
     parser.add_argument("--start", type=str, default='2024-03-01')
     parser.add_argument("--stop", type=str, default=now)
     parser.add_argument("--monthly", action='store_true')
+    parser.add_argument("--mode", choices=['append','replace'])
     args = parser.parse_args()
 
     exec_query(args.table, args.db_origin, args.db_target, args.start, args.stop, args.monthly)
